requirements.txt
APScheduler==3.6.3
arrow==1.1.1
asgiref==3.4.1
beautifulsoup4==4.9.3
blessed==1.18.1
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
Django==3.2.5
django-picklefield==3.0.1
idna==3.2
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
wcwidth==0.2.5
